+++ src/sensors/windshield-camera.c
